grpc: reorganize file structure
* merge small files into one.
